+++ src/workflows/workflows.ts
@@ -20,7 +20,12 @@ import {
   registerOffchainWorkflowWithSigner,
 } from './registration';
 import { mintingWorkflow } from './minting';
-import { transfersWorkflow, batchTransfersWorkflow } from './transfers';
+import {
+  transfersWorkflow,
+  batchTransfersWorkflow,
+  transfersWorkflowWithSigner,
+  batchTransfersWorkflowWithSigner,
+} from './transfers';
 import {
   depositERC20Workflow,
   depositERC721Workflow,
@@ -113,6 +118,7 @@ export class Workflows {
     return mintingWorkflow(signer, request, this.mintsApi);
   }
 
+  /** @deprecated */
   public transfer(
     signer: Signer,
     starkWallet: StarkWallet,
@@ -121,6 +127,12 @@ export class Workflows {
     return transfersWorkflow(signer, starkWallet, request, this.transfersApi);
   }
 
+  public transferWithSigner(walletConnection: WalletConnection, request: UnsignedTransferRequest) {
+    return transfersWorkflowWithSigner(
+      { ...walletConnection, request, transfersApi: this.transfersApi });
+  }
+
+  /** @deprecated */
   public batchNftTransfer(
     signer: Signer,
     starkWallet: StarkWallet,
@@ -131,6 +143,17 @@ export class Workflows {
       starkWallet,
       request,
       this.transfersApi,
+    );
+  }
+
+  public batchNftTransferWithSigner(
+    walletConnection: WalletConnection, request: UnsignedBatchNftTransferRequest,
+  ) {
+    return batchTransfersWorkflowWithSigner({
+      ...walletConnection,
+      request,
+      transfersApi: this.transfersApi,
+    },
     );
   }